Start implementing mixin operators.
Also removes the requirement of having all rules in an Actions
context.
