+++ nengo_spa/modules/compare.py
@@ -2,7 +2,7 @@ import numpy as np
 
 import nengo
 from nengo.params import Default, IntParam
-from nengo_spa.network import Network
+from nengo_spa.ast import Network
 from nengo_spa.vocab import VocabularyOrDimParam
