+++ nengo_spa/modules/state.py
@@ -2,7 +2,7 @@ import nengo
 from nengo.exceptions import ValidationError
 from nengo.networks.ensemblearray import EnsembleArray
 from nengo.params import BoolParam, Default, IntParam, NumberParam
-from nengo_spa.network import Network
+from nengo_spa.ast import Network
 from nengo_spa.vocab import VocabularyOrDimParam
 from nengo_spa.networks import IdentityEnsembleArray
